Feature Selection with Confusion Matrix
This code performs logistic regression on a diabetes dataset to predict the onset of diabetes based on three different features: pregnancy, blood pressure, and insulin.
For each of the three chosen features, logistic regression is performed using the sm.add_constant and LogisticRegression functions from the statsmodels and sklearn libraries, respectively. The model is trained on the training data and then used to predict the target variable on the testing data. The confusion matrix and classification report are then printed to evaluate the accuracy of the logistic regression model for each feature.